Changed to powerline symbols for consistency.
tmux-powerline, by default, uses the symbols in place for vim-powerline
rather than the symbols which have been put in place for powerline.

Since vim-powerline will not be supported going forward, (and since I
use powerline in vim) it doesn't make sense to continue to use its
symbols. By aligning the symbols, it becomes much easier to patch a font
for use with both powerline tools (until the tools become aligned).